add assert in Parser.java
Parser class has objects that might be not initialize and fields that will never reach.

Let's add asserts to ensure the variable is correctly initialize and the field is never reach.

Using assertions can avoid error caused by programmers and to make sure the invariant is correct at that point of code.