Refactor: refactor javaDocs and error messages
all the method description of the javadoc does not have an empty line to separate it from other parameters, decreases readability. There are also confusing error message that needs to be changed.

Let's put an empty line to separate it and edit some error messages.

This way can greatly increase readability and user experience for the error message output.